Added a ton of permissions and cleaned up schema

+++ pages/api/graphql/schemas/Character/resolvers/Mutation/deleteCharacter.ts
@@ -2,8 +2,24 @@ import { logger } from '$helpers-server';
 import { Mutation, MutationDeleteProjectArgs, ResolverFn } from '$types';
 import { ApolloError } from 'apollo-server-micro';
 
-export const deleteCharacter: ResolverFn<MutationDeleteProjectArgs, Promise<Mutation['deleteProject']>> = async (_, { id }, { prisma }) => {
+export const deleteCharacter: ResolverFn<MutationDeleteProjectArgs, Promise<Mutation['deleteProject']>> = async (
+  _,
+  { id },
+  { prisma, unpackedToken },
+) => {
   try {
+    const userCharacters = await prisma.character.findMany({
+      where: {
+        userId: unpackedToken.id,
+      },
+    });
+
+    const characterIds = userCharacters.map((character) => character.id);
+
+    if (!characterIds.includes(id)) {
+      throw new ApolloError('You do not have permission to delete this character');
+    }
+
     await prisma.project.delete({
       where: {
         id,
