+++ pages/api/graphql/schemas/Project/resolvers/Mutation/deleteProject.ts
@@ -2,8 +2,27 @@ import { logger } from '$helpers-server';
 import { Mutation, MutationDeleteProjectArgs, ResolverFn } from '$types';
 import { ApolloError } from 'apollo-server-micro';
 
-export const deleteProject: ResolverFn<MutationDeleteProjectArgs, Promise<Mutation['deleteProject']>> = async (_, { id }, { prisma }) => {
+export const deleteProject: ResolverFn<MutationDeleteProjectArgs, Promise<Mutation['deleteProject']>> = async (
+  _,
+  { id },
+  { prisma, unpackedToken },
+) => {
   try {
+    const ownedProjects = await prisma.user.findUnique({
+      where: {
+        id: unpackedToken.id,
+      },
+      select: {
+        ownedProjects: true,
+      },
+    });
+
+    const ownedProjectIds = ownedProjects?.ownedProjects.map((project) => project.id) ?? [];
+
+    if (!ownedProjectIds.includes(id)) {
+      throw new ApolloError('You do not have permission to delete this project');
+    }
+
     await prisma.project.delete({
       where: {
         id,
